refactor: RoomModule을 기반으로 RoomRepository에서 REDIS_CLIENT를 바로 받아 초기화 진행하도록 수
