+++ server/src/app.module.ts
@@ -3,8 +3,9 @@ import { AppController } from '@/app.controller';
 import { AppService } from '@/app.service';
 import { CommonModule } from '@/common/common.module';
 import { ConfigModule } from '@nestjs/config';
-import { RedisModule } from '@/common/redis/redis.module';
-import { RoomRepository } from './room/repository/room.repository';
+import { REDIS_CLIENT, RedisModule } from '@/common/redis/redis.module';
+import { RoomRepository } from './room/room.repository';
+import { RedisClientType } from 'redis';
 
 @Module({
   imports: [CommonModule, ConfigModule.forRoot(), RedisModule],
